#481 fixed spme failures
