+++ pybamm/discretisations/discretisation.py
@@ -429,7 +429,10 @@ class Discretisation(object):
             # note we are sending in the key.id here so we don't have to
             # keep calling .id
             pybamm.logger.debug("Discretise {!r}".format(eqn_key))
-            new_var_eqn_dict[eqn_key] = self.process_symbol(eqn)
+            try:
+                new_var_eqn_dict[eqn_key] = self.process_symbol(eqn)
+            except:
+                new_var_eqn_dict[eqn_key] = self.process_symbol(eqn)
 
         return new_var_eqn_dict
