+++ pybamm/expression_tree/binary_operators.py
@@ -406,9 +406,15 @@ class MatrixMultiplication(BinaryOperator):
         # is a (slice of a) state vector, e.g. for discretised spatial
         # operators of the form D @ u
         left, right = self.orphans
-        if isinstance(left, pybamm.Array):
+        if isinstance(left, pybamm.Array) or (
+            isinstance(left, pybamm.Negate) and isinstance(left.child, pybamm.Array)
+        ):
             left = pybamm.Matrix(csr_matrix(left.evaluate()))
-            return left @ right.jac(variable)
+            jac = left @ right.jac(variable)
+            # Remove domain as matrix multiplies which come from an integral
+            # end up getting the child domains back which throws an error
+            jac.domain = []
+            return jac
         else:
             raise NotImplementedError
