+++ pybamm/models/submodels/electrolyte/stefan_maxwell/diffusion/base_stefan_maxwell_diffusion.py
@@ -2,6 +2,7 @@
 # Base class for electrolyte diffusion employing stefan-maxwell
 #
 from ...base_electrolyte_diffusion import BaseElectrolyteDiffusion
+import pybamm
 
 
 class BaseModel(BaseElectrolyteDiffusion):
@@ -19,3 +20,14 @@ class BaseModel(BaseElectrolyteDiffusion):
 
     def __init__(self, param):
         super().__init__(param)
+
+    def set_boundary_conditions(self, variables):
+
+        c_e = variables["Electrolyte concentration"]
+
+        self.boundary_conditions = {
+            c_e: {
+                "left": (pybamm.Scalar(0), "Neumann"),
+                "right": (pybamm.Scalar(0), "Neumann"),
+            }
+        }
